added screen

+++ Model/Attachement.cs
@@ -3,7 +3,7 @@
     public class Attachement
     {
         public int Id { get; set; }
-        public int MailId { get; set; }
+        public long MailId { get; set; }
         public string FileName { get; set; }
         public string FileExtention { get; set; }
         public string Path { get; set; } //where it should be saved
